restructuring of ChartRenderer file: created types.ts, EventHandler.ts and ChartOptions.ts
