+++ packages/suite-base/src/panels/Plot/ChartRenderer.ts
@@ -5,117 +5,32 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-import { Chart, ChartDataset, ChartOptions, ScatterDataPoint } from "chart.js";
+import { Chart } from "chart.js";
 import { AnnotationOptions } from "chartjs-plugin-annotation";
 import EventEmitter from "eventemitter3";
 
 import { Zoom as ZoomPlugin } from "@lichtblick/chartjs-plugin-zoom";
 import { unwrap } from "@lichtblick/den/monads";
 import { Immutable } from "@lichtblick/suite";
-import { Bounds, Bounds1D } from "@lichtblick/suite-base/types/Bounds";
+import { getChartOptions } from "@lichtblick/suite-base/panels/Plot/ChartUtilities/ChartOptions";
+import {
+  addEventListener,
+  removeEventListener,
+} from "@lichtblick/suite-base/panels/Plot/ChartUtilities/EventHandler";
+import { Bounds } from "@lichtblick/suite-base/types/Bounds";
 import { maybeCast } from "@lichtblick/suite-base/util/maybeCast";
-import { fontMonospace } from "@lichtblick/theme";
 
-import { OriginalValue } from "./datum";
-
-export type Scale = {
-  min: number;
-  max: number;
-  left: number;
-  right: number;
-};
-
-type BaseInteractionEvent = {
-  cancelable: boolean;
-  deltaY: number;
-  deltaX: number;
-
-  boundingClientRect: DOMRect;
-};
-
-type MouseBase = BaseInteractionEvent & {
-  clientX: number;
-  clientY: number;
-};
-
-type WheelInteractionEvent = { type: "wheel" } & BaseInteractionEvent & MouseBase;
-type PanStartInteractionEvent = { type: "panstart" } & BaseInteractionEvent & {
-    center: { x: number; y: number };
-  };
-type PanMoveInteractionEvent = { type: "panmove" } & BaseInteractionEvent;
-type PanEndInteractionEvent = { type: "panend" } & BaseInteractionEvent;
-
-export type InteractionEvent =
-  | WheelInteractionEvent
-  | PanStartInteractionEvent
-  | PanMoveInteractionEvent
-  | PanEndInteractionEvent;
-
-export type Datum = ScatterDataPoint & { value?: OriginalValue };
-export type Dataset = ChartDataset<"scatter", Datum[]>;
-
-type ChartType = Chart<"scatter", Datum[]>;
-
-export type HoverElement = {
-  data: Datum;
-  configIndex: number;
-};
-
-export type Size = { width: number; height: number };
-
-export type UpdateAction = {
-  type: "update";
-  size?: { width: number; height: number };
-  showXAxisLabels?: boolean;
-  showYAxisLabels?: boolean;
-  xBounds?: Partial<Bounds1D>;
-  yBounds?: Partial<Bounds1D>;
-  zoomMode?: "x" | "y" | "xy";
-  referenceLines?: { color: string; value: number }[];
-  interactionEvents?: InteractionEvent[];
-};
-
-function addEventListener(emitter: EventEmitter) {
-  return (eventName: string, fn?: () => void) => {
-    const existing = emitter.listeners(eventName);
-    if (!fn || existing.includes(fn)) {
-      return;
-    }
-
-    emitter.on(eventName, fn);
-  };
-}
-
-function removeEventListener(emitter: EventEmitter) {
-  return (eventName: string, fn?: () => void) => {
-    if (fn) {
-      emitter.off(eventName, fn);
-    }
-  };
-}
-
-// allows us to override the chart.ctx instance field which zoom plugin uses for adding event listeners
-type MutableContext<T> = Omit<Chart, "ctx"> & { ctx: T };
-
-type PanEvent = {
-  deltaX: number;
-  deltaY: number;
-};
-
-type PanStartEvent = PanEvent & {
-  center: { x: number; y: number };
-  target: {
-    getBoundingClientRect(): DOMRect;
-  };
-};
-
-type ZoomableChart = Chart & {
-  $zoom: {
-    panStartHandler(event: PanStartEvent): void;
-    panHandler(event: PanEvent): void;
-    panEndHandler(): void;
-  };
-};
+import {
+  ChartType,
+  Dataset,
+  Datum,
+  HoverElement,
+  InteractionEvent,
+  MutableContext,
+  Scale,
+  UpdateAction,
+  ZoomableChart,
+} from "./types";
 
 export class ChartRenderer {
   #chartInstance: ChartType;
@@ -137,6 +52,12 @@ export class ChartRenderer {
       },
     };
 
+    const chartOptions = getChartOptions({
+      devicePixelRatio: args.devicePixelRatio,
+      gridColor: args.gridColor,
+      tickColor: args.tickColor,
+    });
+
     const origZoomStart = ZoomPlugin.start?.bind(ZoomPlugin);
     ZoomPlugin.start = (chartInstance: MutableContext<unknown>, startArgs, pluginOptions) => {
       // swap the canvas with our fake dom node canvas to support zoom plugin addEventListener
@@ -149,74 +70,6 @@ export class ChartRenderer {
       return res;
     };
 
-    const fullOptions: ChartOptions<"scatter"> = {
-      maintainAspectRatio: false,
-      animation: false,
-      // Disable splines, they seem to cause weird rendering artifacts:
-      elements: { line: { tension: 0 } },
-      interaction: {
-        intersect: false,
-        mode: "x",
-      },
-      devicePixelRatio: args.devicePixelRatio,
-      font: { family: fontMonospace, size: 10 },
-      // we force responsive off since we manually trigger width/height updates on the chart
-      // responsive mode does not work properly with offscreen canvases and retina device pixel ratios
-      // it results in a run-away canvas that keeps doubling in size!
-      responsive: false,
-      scales: {
-        x: {
-          type: "linear",
-          display: true,
-          grid: { color: args.gridColor },
-          ticks: {
-            font: {
-              family: fontMonospace,
-              size: 10,
-            },
-            color: args.tickColor,
-            maxRotation: 0,
-          },
-        },
-        y: {
-          type: "linear",
-          display: true,
-          grid: { color: args.gridColor },
-          ticks: {
-            font: {
-              family: fontMonospace,
-              size: 10,
-            },
-            color: args.tickColor,
-            padding: 0,
-            precision: 3,
-          },
-        },
-      },
-      plugins: {
-        decimation: {
-          enabled: false,
-        },
-        tooltip: {
-          enabled: false, // Disable native tooltips since we use custom ones.
-        },
-        zoom: {
-          zoom: {
-            enabled: true,
-            mode: "x",
-            sensitivity: 3,
-            speed: 0.1,
-          },
-          pan: {
-            mode: "xy",
-            enabled: true,
-            speed: 20,
-            threshold: 10,
-          },
-        },
-      },
-    };
-
     // ChartJS supports offscreen canvas however the type definitions do not so we need to cast and
     // fool the constructor.
     //
@@ -227,7 +80,7 @@ export class ChartRenderer {
       data: {
         datasets: [],
       },
-      options: fullOptions,
+      options: chartOptions,
       plugins: [ZoomPlugin],
     });
